rename module to bounded_deque
